update page tab bar
