+++ frontend/src/pages/termsOfUse/TermsOfUse.tsx
@@ -1,31 +1,26 @@
 import React from 'react';
-import {
-  Unstable_Grid2 as Grid
-} from '@mui/material';
 
 import BasePage from '../../pages/BasePage';
-import TabBar from 'components/TabBar';
+import PageTabBar from '../../components/PageTabBar';
 import ForAdults from './ForAdults/ForAdults';
 import ForChildren from './ForChildren/ForChildren';
 
 const TermsOfUse: React.FC = () => {
   return (
     <BasePage>
-      <Grid xs={12}>
-        <TabBar
-          title='Terms of use'
-          tabs={[
-            {
-              label: 'Terms of use',
-              element: <ForAdults />
-            },
-            {
-              label: 'Child-friendly',
-              element: <ForChildren />
-            }
-          ]}
-        />
-      </Grid>
+      <PageTabBar
+        title='Terms of use'
+        tabs={[
+          {
+            label: 'Terms of use',
+            element: <ForAdults />
+          },
+          {
+            label: 'Child-friendly',
+            element: <ForChildren />
+          }
+        ]}
+      />
     </BasePage>
   );
 };
